courseapis/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cloudinary==1.42.2
Django==5.1.6
django-ckeditor==6.7.2
django-js-asset==3.0.1
djangorestframework==3.15.2
drf-yasg==1.21.9
inflection==0.5.1
packaging==24.2
pillow==11.1.0
PyMySQL==1.1.1
pytz==2025.1
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0